store activated interfaces in config object
On a new installation, they don't get saved in the activated_plugins list (and
thus not moved up in the acl_users plugin registry list, causing #7).  (Import
was working properly already.)  This fixes #7 properly and completely for both
AD and non-AD, although we may wish to also move up other plugins, e.g.
IGroupManagement, which needs to be at the top for modifying group membership to
work.

Also removed the confusingly-named 'activated_interfaces' variable, which was
never used anywhere ('activated_plugins' is the one actually used).