add mandatory schema items to local config object
Reorder/modify other things so that the local schema object is configured
first, then the whole thing is pushed to the LDAPUserFolder.

Fixes the bug where adding a new AD configuration would have the schema
items stored in the LUF but not config.schema, so they would only show up
in the ZMI, not Plone.