don't allow removing required attributes
Also adjust the docs to match and clean up/clarify them.

NB: removed the callback from HandleRemoved(), because the property is not
yet removed from config.schema at this point, and it was getting added back
to the LUF.

Preventing deletion of a mandatory property is a better solution than deleting
it and then immediately adding it back.