Import & export LDAP group to Plone role mappings
Because these cannot currently[1] be set via Plone (NotImplemented exception
in Products.PloneLDAP.mixins.groupmgmt), these must be set via the
LDAPUserFolder in the ZMI, and should be saved/restored with the rest of the
LDAP settings.

These are stored as the '_groups_mapping' dict in the LDAPUserFolder object;
it's easiest to export this as a string and parse it with ast during import.

[1] Fixed in collective/Products.PloneLDAP@9a90f98, but no new release has
been cut yet.