Added dependencies so Plone 5 can start.
Note that these dependencies {zope,five}.formlib and plone.app.form,
are in Plon 4.2 and 4.3 by default, so nothing changes there.

Removed the Plone 5 classifiers, because there is a test failure,
and it can't have worked before.

Also removed the Python 2.6 classifier, because we really do not test with that anymore.